backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
mysqlclient==2.2.4
django-cors-headers==4.4.0
gunicorn==21.2.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2